pyproject.toml
[tool.poetry]
name = "ironbot"
version = "0.0.6"
description = "CLI to get information about Ironman professional races"
authors = ["Eduardo Cuducos <4732915+cuducos@users.noreply.github.com>"]
license = "GPLv3"
readme = "README.md"
repository = "https://github.com/cuducos/ironbot/"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Utilities",
    ]
keywords = [
    "triathlon",
    "Ironamn",
    "Professional triathletes",
    "Professional triathlon races",
    ]
packages = [
     { include = "ironbot" },
    ]

[tool.poetry.scripts]
ironbot = "ironbot.__main__:app"

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.11.1"
beautifulsoup4 = "^4.12.2"
camelot-py = "^0.11.0"
httpx = "^0.24.1"
sqlalchemy = "^2.0.15"
typer = "^0.9.0"

# we could use camelot-py[base] but that would inject problematic pdftopng, so
# let's add the other base packages manually. See, for example:
# https://github.com/vinayak-mehta/pdftopng/issues/11
# https://github.com/vinayak-mehta/pdftopng/issues/8
ghostscript = "^0.7"
opencv-python = "^4.7.0.72"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^8.13.2"
pytest = "^7.3.1"
pytest-black = "^0.3.12"
pytest-mypy = "^0.10.3"
pytest-ruff = "^0.1"
pytest-watch = "^4.2.0"
types-beautifulsoup4 = "^4.12.0.5"

[tool.pytest.ini_options]
addopts = "--black --mypy --ruff"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
